==> code/2.3_lit_numbers.R <==
# Smith et al. 2019 

sum(c(19426, 29016, 5301, 5281)) * 10 # usfs forest in km2

sum(c(23194, 38878, 6472, 8818))*10 # fed forest in km2

sum(c(19426, 29016, 5301, 5281)) / sum(c(23194, 38878, 6472, 8818))


13761/17212 #frac of federal forest ecosystem C found in national forests

112290/ sum(c(19426, 29016, 5301, 5281)) * 10 # fraction of usfs forest that is 'reserved'
